Fixed errors and refactored code

+++ src/main/java/com/spring/controllers/ApiProductController.java
@@ -61,7 +61,7 @@ public class ApiProductController {
     @GetMapping(path = "/add")
     public Product addProduct(@RequestParam String name, @RequestParam String description, @RequestParam(required = false) byte[] bytes,
                               @RequestParam String brand, @RequestParam String material, @RequestParam String manufacturerCountry,
-                              @RequestParam String category, @RequestParam int price, @RequestParam double weight) {
+                              @RequestParam String category, @RequestParam double price, @RequestParam double weight) {
         if (!name.isEmpty() && !description.isEmpty() && !material.isEmpty() && !brand.isEmpty() &&
                 !manufacturerCountry.isEmpty() && !category.isEmpty() && price >= 0 && weight > 0)
             return productService.addProduct(new Product(name, description, bytes, brand, material,
@@ -89,7 +89,7 @@ public class ApiProductController {
     public Product updateProduct(@PathVariable("id") int vendorId, @RequestParam String name,
                                  @RequestParam String description, @RequestParam(required = false) byte[] bytes, @RequestParam String brand,
                                  @RequestParam String material, @RequestParam String manufacturerCountry,
-                                 @RequestParam String category, @RequestParam int price, @RequestParam float weight) {
+                                 @RequestParam String category, @RequestParam double price, @RequestParam float weight) {
         if (!name.isEmpty() && !description.isEmpty() && !material.isEmpty() && !brand.isEmpty() &&
                 !manufacturerCountry.isEmpty() && !category.isEmpty() && price >= 0 && weight > 0 && vendorId > 0)
             return productService.updateProduct(vendorId,
